refactor(DEX-4063): new responsive approach
