+++ .env.development
@@ -44,7 +44,7 @@ SUPPORT_URL_CALCULATOR_MATH='https://support.edx.org/hc/en-us/articles/360000038
 SUPPORT_URL_ID_VERIFICATION='https://support.edx.org/hc/en-us/articles/206503858-How-do-I-verify-my-identity'
 SUPPORT_URL_VERIFIED_CERTIFICATE='https://support.edx.org/hc/en-us/articles/206502008-What-is-a-verified-certificate'
 TERMS_OF_SERVICE_URL='https://www.edx.org/edx-terms-service'
-HONOR_CODE_URL='https://cm.wgu.edu/t5/WGU-Student-Policy-Handbook/Student-Code-of-Conduct/ta-p/129'
+HONOR_CODE_URL=''
 USER_INFO_COOKIE_NAME='edx-user-info'
 SESSION_COOKIE_DOMAIN='.local.overhang.io'
